+++ kernel/application/views/formats/specialview/standard.php
@@ -1,0 +1,10 @@
+<?php
+
+// ****** Load Element ******
+$Output .= $this->LoadElement("specialview");
+
+                                                                                     
+// ****** Add Messagebar ******
+$Output .= "<p></p><div id='special_messagebar'></div>";
+
+?>