+++ kernel/application/views/formats/tabs/userfees.php
@@ -23,7 +23,7 @@ $Output .= "<tr>";
 $Output .= "<th width='75px'>" . $this->CI->database->code2text("DATE")        . "</th>";
 if ( $this->countLBS() == 2 ) $Output .= "<th>" . $this->CI->database->code2text("LIBRARY")   . "</th>";
 $Output .= "<th>" . $this->CI->database->code2text("DESCRIPTION") . "</th>";
-$Output .= "<th>" . $this->CI->database->code2text("AMOUNT")      . "</th>";
+$Output .= "<td align='right'><b>" . $this->CI->database->code2text("AMOUNT")      . "</b></td>";
 $Output .= "</tr>";
 
 // Print data
@@ -35,7 +35,14 @@ foreach ( $Items as $Item )
   if ( $this->countLBS() == 2 ) $Output .= "<td>" .$this->getLBSName($Item["isil"])   . "</td>";
   $Output .= "<td>";
   if ( isset($Item["feetype"]) ) $Output .= $Item["feetype"];
-  if ( isset($Item["about"]) ) $Output .= "<br />" . $Item["about"];
+  if ( isset($Item["about"])  ) 
+  {
+    if ( !isset($Item["feetype"]) || ( isset($Item["feetype"]) && $Item["feetype"] != $Item["about"] ) )
+    {
+      if ( isset($Item["feetype"]) )  $Output .= "<br />";
+      $Output .= $Item["about"];
+    }
+  }
   $Output .= "</td><td align='right'>";
   if ( isset($Item["amount"]) )
   {
@@ -52,11 +59,11 @@ if ( isset($_SESSION["info"]["1"]["isil"]) && isset($_SESSION[$_SESSION["info"][
 	$Output .= "<tr>";
   $Output .= "<td></td>";
   if ( $this->countLBS() == 2 ) $Output .= "<td></td>";
-	$Output .= "<td align='right'>" . $this->CI->database->code2text("FEETOTAL") . "</td>";
-	$Output .= "<td width='80px' align='right'>";
+	$Output .= "<td align='right'><b>" . $this->CI->database->code2text("FEETOTAL") . "</b></td>";
+	$Output .= "<td width='80px' align='right'><b>";
   $Wert    = (float) explode(" ",$_SESSION[$_SESSION["info"]["1"]["isil"]]["fees"]["amount"])[0];
   $Output .= $this->formatEuro($Wert);
-	$Output .= "</td></tr>";
+	$Output .= "</b></td></tr>";
 }
 
 ?>